Attempt to remove warnings v4

+++ distributedcombigrid/src/sgpp/distributedcombigrid/fullgrid/DistributedFullGrid.hpp
@@ -1397,7 +1397,7 @@ class DistributedFullGrid {
       // check if process grid of comm uses the required ordering
       auto maxdims = procs_.size();
       std::vector<int> cartdims(maxdims), periods(maxdims), coords(maxdims);
-      MPI_Cart_get(comm, maxdims, &cartdims[0], &periods[0], &coords[0]);
+      MPI_Cart_get(comm, static_cast<int>(maxdims), &cartdims[0], &periods[0], &coords[0]);
 
       assert(cartdims == dims);
